Import pico-bitcoin-wallet files
Import Kixunil's workshop but from tcharding's fork with minor
improvements. Done like this instead of PR'ing because it is within the
spirit of the workshop repo, which is just to hold code that can be used
as wanted to hold workshops. We aren't precious about this code or
changes to it.